+++ calibrate.h
@@ -1,6 +1,7 @@
+#include "stddef.h"
 
 void CalInit(void);
-bool CalRead(void *data, int len);
-bool CalWrite(void *data, int len);
+bool CalRead(void *data, size_t len);
+bool CalWrite(void *data, size_t len);
